rename: App에 있던 router 코드를 Router.js 파일로 이동 (#33)
