+++ src/App.jsx
@@ -1,28 +1,9 @@
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import {
-	ErrorPage,
-	LandingPage,
-	LoginPage,
-	PersonalSchedulePage,
-	Root,
-	SignUpPage,
-} from "@/pages";
+import { RouterProvider } from "react-router-dom";
 import { useDispatch } from "react-redux";
 import { useEffect } from "react";
 import { getCurrentUser } from "./store/user/user-slice";
 import { ToastContainer } from "react-toastify";
-
-const router = createBrowserRouter([
-	{
-		path: "/",
-		element: <Root />,
-		errorElement: <ErrorPage />,
-		children: [{ index: true, element: <PersonalSchedulePage /> }],
-	},
-	{ path: "/landing", element: <LandingPage /> },
-	{ path: "/login", element: <LoginPage /> },
-	{ path: "/signup", element: <SignUpPage /> },
-]);
+import router from "./Router";
 
 export default function App() {
 	const dispatchFn = useDispatch();
